API/requirements.txt
tensorflow-cpu
flask
numpy
pandas
matplotlib
plotly
seaborn
scikit-learn
opencv-python
gunicorn